refactor: mypage
