+++ src/app/profile/[id]/page.tsx
@@ -6,9 +6,9 @@ import { useGetMembersById, useGetMembersMe } from '@/apis/member';
 import { useGetMissionStack } from '@/apis/mission';
 import { FollowStatus } from '@/apis/schema/member';
 import FollowButton from '@/app/profile/[id]/FollowButton';
-import ProfileContent from '@/app/profile/[id]/ProfileContent';
 import BottomDim from '@/components/BottomDim/BottomDim';
 import Header from '@/components/Header/Header';
+import ProfileContent from '@/components/ProfileContent';
 import ProfileTab from '@/components/ProfileTab/ProfileTab';
 import { ROUTER } from '@/constants/router';
 import { css } from '@styled-system/css';
