+++ src/components/ProfileTab/ProfileTab.tsx
@@ -1,13 +1,13 @@
-'use client';
 import { useGetMembersMe } from '@/apis/member';
-import ProfileFeedList from '@/components/ProfileTab/ProfileFeedList';
-import ProfileMissionList from '@/components/ProfileTab/ProfileMissionList';
 import Tab from '@/components/Tab/Tab';
 import { useTab } from '@/components/Tab/Tab.hooks';
 import { type TabType } from '@/components/Tab/Tab.types';
 import { EVENT_LOG_CATEGORY, EVENT_LOG_NAME } from '@/constants/eventLog';
 import { eventLogger } from '@/utils';
 import { css } from '@styled-system/css';
+
+import ProfileFeedList from './ProfileFeedList';
+import ProfileMissionList from './ProfileMissionList';
 
 const PROFILE_TABS = [
   {
@@ -20,8 +20,12 @@ const PROFILE_TABS = [
   },
 ];
 
+interface ProfileTabProps {
+  memberId: number;
+}
+
 // TODO : mypage, profile에서 따로 사용하도록 리팩토링 하면 좋을 듯
-function ProfileTab({ memberId }: { memberId: number }) {
+function ProfileTab({ memberId }: ProfileTabProps) {
   const { tabs, activeTab, onTabClick } = useTab(PROFILE_TABS, 'feed');
   const { data } = useGetMembersMe();
   if (!data) return null;
@@ -48,7 +52,13 @@ function ProfileTab({ memberId }: { memberId: number }) {
 
 export default ProfileTab;
 
-function TabContent({ activeTab, memberId, isMySelf }: { activeTab: string; memberId: number; isMySelf: boolean }) {
+interface TabContentProps {
+  activeTab: string;
+  memberId: number;
+  isMySelf: boolean;
+}
+
+function TabContent({ activeTab, memberId, isMySelf }: TabContentProps) {
   if (activeTab === 'feed') {
     return <ProfileFeedList memberId={memberId} isMySelf={isMySelf} />;
   }
